#847 Markdown interface
